# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hybrid Cloud Works Migration Explorer API (web-front edition)
  version: 0.3.0
  description: |
    CSV-based assessment API behind the migration AddOn pane (`https://migration.lab.hybridcloudworks.com`). Never connects
    to Azure, never accepts credentials, never executes operations. Assessment routes require the `x-owner-token` header
    issued at creation; assessments expire (default 120 min) and live in memory only.

    Every response carries `X-Addon-Id: migration`, `X-Addon-Version`, `X-Content-Type-Options: nosniff`,
    `Referrer-Policy: no-referrer`, a Content Security Policy whose `frame-ancestors` comes from `AMO_FRAME_ANCESTORS`
    and, for API responses, `Cache-Control: no-store`. Errors use one envelope: `{ error: { code, message, details? } }`.
servers:
  - url: /
paths:
...
  /api/stats:
    get:
      summary: Opt-in aggregate counts (AMO_TELEMETRY=1); never names, IDs, tags, subscriptions or file contents
      responses:
        "200": { description: "Aggregates since process start", headers: *addon_headers }
        "404": { description: "not_enabled \u2014 telemetry is off (the default)", headers: *addon_headers }
...
  /api/assessments:
    post:
      summary: Create an assessment from a resources.csv export
      description: |
        Anonymous mutating route. Requires a human-verification token (`x-turnstile-token`) whenever `TURNSTILE_SECRET` is
        configured; without a secret the route answers 503 `turnstile_not_configured` unless the deployment explicitly allows
        unverified uploads (local development only). Rate limited per client address (`AMO_RATE_LIMIT_POSTS` per
        `AMO_RATE_LIMIT_WINDOW_MINUTES`) and bounded to `AMO_MAX_CONCURRENT` assessments at once.
      parameters:
        - { name: x-turnstile-token, in: header, required: false, schema: { type: string }, description: Human-verification token from the widget rendered with the site key published by /api/health }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [csv]
              properties:
                csv: { type: string, maxLength: 5242880 }
                fileName: { type: string, pattern: "^[\\w .()-]{1,120}\\.csv$" }
                intent: { $ref: "#/components/schemas/MigrationIntentInput" }
          text/csv:
            schema: { type: string }
      responses:
        "201": { description: "Created", content: { application/json: { schema: { $ref: "#/components/schemas/CreateAssessmentResponse" } } }, headers: *addon_headers }
        "400": { description: "Invalid body, prohibited credential-shaped intent field, or bad file name", content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }, headers: *addon_headers }
        "403": { description: "turnstile_failed \u2014 human verification missing or rejected", content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }, headers: *addon_headers }
        "413": { description: "too_large \u2014 body exceeds 5 MB", headers: *addon_headers }
        "422": { description: "ingestion_failed \u2014 CSV could not be ingested (missing NAME/TYPE columns, empty, binary)", headers: *addon_headers }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
...
  headers:
...
    XAddonId: { description: "x-addon-id: the catalogue id", schema: { type: string, const: migration } }
    XAddonVersion: { description: "x-addon-version: semver of the running image", schema: { type: string } }
    XContentTypeOptions: { description: "x-content-type-options", schema: { type: string, const: nosniff } }
    ReferrerPolicy: { description: "referrer-policy", schema: { type: string, const: no-referrer } }
    PermissionsPolicy: { description: "permissions-policy", schema: { type: string, const: "camera=(), microphone=(), geolocation=()" } }
    CrossOriginOpenerPolicy: { description: "cross-origin-opener-policy", schema: { type: string, const: same-origin } }
  responses:
    RateLimited:
      description: rate_limited — the client's token bucket is empty; `Retry-After` (seconds) says when a token is back
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until the next request may succeed }
        X-Addon-Id: { $ref: "#/components/headers/XAddonId" }
        X-Addon-Version: { $ref: "#/components/headers/XAddonVersion" }
        X-Content-Type-Options: { $ref: "#/components/headers/XContentTypeOptions" }
        Referrer-Policy: { $ref: "#/components/headers/ReferrerPolicy" }
        Permissions-Policy: { $ref: "#/components/headers/PermissionsPolicy" }
        Cross-Origin-Opener-Policy: { $ref: "#/components/headers/CrossOriginOpenerPolicy" }
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    Unavailable:
      description: |
        `overloaded` — more than `AMO_MAX_CONCURRENT` assessments are running (`Retry-After: 5`), or
        `turnstile_not_configured` — human verification is not configured on this deployment and unverified uploads are not allowed.
      headers:
        Retry-After: { schema: { type: integer }, description: Present for overloaded }
        X-Addon-Id: { $ref: "#/components/headers/XAddonId" }
        X-Addon-Version: { $ref: "#/components/headers/XAddonVersion" }
        X-Content-Type-Options: { $ref: "#/components/headers/XContentTypeOptions" }
        Referrer-Policy: { $ref: "#/components/headers/ReferrerPolicy" }
        Permissions-Policy: { $ref: "#/components/headers/PermissionsPolicy" }
        Cross-Origin-Opener-Policy: { $ref: "#/components/headers/CrossOriginOpenerPolicy" }
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Stable snake_case machine code (rate_limited, overloaded, turnstile_not_configured, turnstile_failed, too_large, invalid_json, ingestion_failed, not_found, not_configured, not_enabled, execution_disabled, internal) }
            message: { type: string, description: Visitor-safe; no hostnames, paths or stacks }
            details: { description: Optional machine detail (e.g. retryAfterSeconds, verification codes, ingestion warnings) }
...
    MigrationIntentInput:
      type: object
      description: Questionnaire answers. Keys containing password/secret/token/key/connectionstring/credential/certificate are rejected.
      properties:
        desiredOperation: { enum: [resource-group-move, subscription-move, region-relocation, cross-tenant-migration, modernization, disaster-recovery] }
        destinationRegion: { type: string }
        sameSubscription: { type: boolean }
        sameTenant: { type: boolean }
        existingLandingZone: { type: boolean }
        environmentType: { enum: [production, non-production, mixed] }
        downtimeTolerance: { enum: [none, minutes, hours, days] }
        rtoHours: { type: number }
        rpoMinutes: { type: number }
        migrationMode: { enum: [online, offline, either] }
        sourceAvailableDuringSync: { type: boolean }
        dataResidency: { type: string }
        regulatoryRestrictions: { type: array, items: { type: string } }
        namingPrefix: { type: string }
        requiredTags: { type: object, additionalProperties: { type: string } }
        applicationGroupTagKey: { type: string }
        publicAccessPolicy: { enum: [allow, deny, unknown] }
        privateEndpointsRequired: { type: boolean }
        includeDataMigrationExamples: { type: boolean }
    CreateAssessmentResponse:
      type: object
      properties:
        assessmentId: { type: string, format: uuid }
        ownerToken: { type: string, description: Shown once; required for every later call }
        expiresAt: { type: string, format: date-time }
        summary: { type: object }
        progress: { type: array, items: { type: object, properties: { at: { type: string }, agent: { type: string }, message: { type: string }, status: { enum: [started, completed, skipped, failed] } } } }
        ingestionWarnings: { type: array, items: { type: string } }